// front/src/api/actions.ts
export enum ACTIONS {
  LOAD_COUNTRIES = "LOAD_COUNTRIES",
  LOAD_RESTRICTIONS = "LOAD_RESTRICTIONS",
  LOAD_SOURCE_DOCUMENT_ORIGINS = "LOAD_SOURCE_DOCUMENT_ORIGINS",
  LOAD_COUNTER_SANCTIONS_RESTRICTIONS = "LOAD_COUNTER_SANCTIONS_RESTRICTIONS",
  LOAD_COUNTER_SANCTIONS_SOURCE_DOCUMENTS = "LOAD_COUNTER_SANCTIONS_SOURCE_DOCUMENTS",
  CHECK_FILTERS = "CHECK_FILTERS",
  SEARCH = "SEARCH",
  SEARCH_COUNTER_SANCTIONS = "SEARCH_COUNTER_SANCTIONS",
  RESET_SANCTIONS_DB_WITH_FILE = "RESET_SANCTIONS_DB_WITH_FILE",
  RESET_COUNTER_SANCTIONS_DB_WITH_FILE = "RESET_COUNTER_SANCTIONS_DB_WITH_FILE",
  AUTH = "AUTH",
  REGISTER = "REGISTER",
  VERIFY = "VERIFY",
  SUBMIT_CONTACT_FORM = "SUBMIT_CONTACT_FORM",
  RECOVER_PASSWORD_REQUEST = "RECOVER_PASSWORD_REQUEST",
  RECOVER_PASSWORD_CONFIRM = "RECOVER_PASSWORD_CONFIRM",
  REGISTRATION_CONFIRM = "REGISTRATION_CONFIRM",
  RETRY_EMAIL_CONFIRM = "RETRY_EMAIL_CONFIRM",
  LOAD_PROFILE = "LOAD_PROFILE",
  CHANGE_PASSWORD = "CHANGE_PASSWORD",
  CHANGE_PROFILE = "CHANGE_PROFILE",
  BILLING_CREATE_PAYMENT = "BILLING_CREATE_PAYMENT",
  BILLING_TARRIF_UPDATED = "BILLING_TARRIF_UPDATED",
  BILLING_TARRIF_USER_NOTICED = "BILLING_TARRIF_USER_NOTICED",
  BILLING_ADD_REQUESTS_PAYMENT = "BILLING_ADD_REQUESTS_PAYMENT",
  SAVE_TARRIFS_SETTINGS = "SAVE_TARRIFS_SETTINGS",
  LOAD_TARRIFS_SETTINGS = "LOAD_TARRIFS_SETTINGS",
  SAVE_COUNTER_SANCTIONS_TARRIFS_SETTINGS = "SAVE_COUNTER_SANCTIONS_TARRIFS_SETTINGS",
  LOAD_COUNTER_SANCTIONS_TARRIFS_SETTINGS = "LOAD_COUNTER_SANCTIONS_TARRIFS_SETTINGS",
  FORCE_LOGOUT = "FORCE_LOGOUT",
  INACTIVITY_LOGOUT = "INACTIVITY_LOGOUT",
  LAST_ACTIVITY_TIME = "LAST_ACTIVITY_TIME",
  LOGOUT = "LOGOUT",
  LOAD_USERS_TARIFFS = "LOAD_USERS_TARIFFS",
  GRANT_TARIFF = "GRANT_TARIFF",
  LOAD_TARRIFS = "LOAD_TARRIFS",
  DELETE_USER_TARIFF = "DELETE_USER_TARIFF",
  GENERATE_EXCEL_REPORT = "GENERATE_EXCEL_REPORT",
  GENERATE_COUNTER_SANCTIONS_EXCEL_REPORT = "GENERATE_COUNTER_SANCTIONS_EXCEL_REPORT",
  REMOVE_REPORT = "REMOVE_REPORT",
  DOWNLOAD_REPORT = "DOWNLOAD_REPORT",
  SAVE_REPORT_TO_MY_REPORTS = "SAVE_REPORT_TO_MY_REPORTS",
  SAVE_COUNTER_SANCTION_REPORT_TO_MY_REPORTS = "SAVE_COUNTER_SANCTION_REPORT_TO_MY_REPORTS",
  LOAD_USER_REPORTS = "LOAD_USER_REPORTS",
  DOWNLOAD_MULTIPLE_REPORTS = "DOWNLOAD_MULTIPLE_REPORTS",
  DELETE_MULTIPLE_REPORTS = "DELETE_MULTIPLE_REPORTS",
  LOAD_PREFERENCES = "LOAD_PREFERENCES",
  DOWNLOAD_EXCEL_TEMPLATE = "DOWNLOAD_EXCEL_TEMPLATE",
  PARSE_SEARCH_EXCEL_FILE = "PARSE_SEARCH_EXCEL_FILE",
}